Add Electron builder config and main entry
